be/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
alembic>=1.11.0
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.3.0
qdrant-client>=1.8.2
google-generativeai>=0.5.2
python-multipart==0.0.20
langchain-core>=0.2.29
langchain-community>=0.3.7
langchain-google-genai>=2.0.7
langgraph>=0.2.35
python-dateutil>=2.9.0.post0
pypdf>=4.2.0
python-docx>=1.1.2
